Improve reconnection logic
Using bluetooth.async_register_callback is not enough when the connection is lost disconnects for a short period of time.
The bluetooth device can be available even for a few minutes after physical device becomes unavailable.
If that happens, and physical device becomes available no discovery callback is triggered.
We have to actively try to connect to the device